+++ server/src/main/java/org/allaymc/server/item/enchantment/EnchantmentOptionGenerator.java
@@ -214,7 +214,7 @@ public final class EnchantmentOptionGenerator {
                 .stream()
                 .filter(item::checkEnchantmentCompatibility)
                 .filter(EnchantmentType::isAvailableInEnchantTable)
-                .filter(type -> type.canBeAppliedTo(item.getItemType()))
+                .filter(type -> item.isAllEnchantmentsAvailableInEnchantTable() || type.canBeAppliedTo(item.getItemType()))
                 .toList();
     }